pyproject.toml
[project]
name = "gemini-inferene-async"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "datasets>=4.2.0",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.35.3",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.1.1",
    "tqdm>=4.67.1",
]
